transus-backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
google-genai==1.0.0
pydantic==2.10.3
pydantic-settings==2.6.1
python-multipart==0.0.17
sqlalchemy==2.0.36
psycopg[binary]==3.2.13
alembic==1.14.0
pytest==8.3.4
httpx==0.27.2

